v0.22.0

+++ app/account/routes.py
@@ -2,29 +2,28 @@ import time
 from datetime import date, datetime, time
 
 import pytz
-from flask import flash, redirect, render_template, request, url_for
-from flask_login import current_user, fresh_login_required, login_required
-
 from app import db, limiter
 from app.account import bp
 from app.account.forms import AccountForm, PasswordForm, ScheduleForm
 from app.main.email import send_confirmation_email
+from flask import flash, redirect, render_template, request, url_for
+from flask_login import current_user, fresh_login_required, login_required
 
 
-@bp.route('/')
+@bp.route("/")
 @login_required
-@limiter.limit("1 per second", key_func=lambda: current_user.id)
+@limiter.limit("2 per second", key_func=lambda: current_user.id)
 def account():
     user = current_user
     user.created_at = user.created_at.astimezone(pytz.timezone(user.timezone))
     user.login_at = user.login_at.astimezone(pytz.timezone(user.timezone))
     user.updated_at = user.updated_at.astimezone(pytz.timezone(user.timezone)) if user.updated_at else None
-    return render_template('account/account.html', title='My Account', user=user)
+    return render_template("account/account.html", title="My Account", user=user)
 
 
-@bp.route('/change-password', methods=['GET', 'POST'])
+@bp.route("/change-password", methods=["GET", "POST"])
 @fresh_login_required
-@limiter.limit("1 per second", key_func=lambda: current_user.id)
+@limiter.limit("2 per second", key_func=lambda: current_user.id)
 def change_password():
     form = PasswordForm()
     if form.validate_on_submit():
@@ -33,19 +32,19 @@ def change_password():
             current_user.updated_at = pytz.utc.localize(datetime.utcnow())
             db.session.add(current_user)
             db.session.commit()
-            flash('Your password has been changed.', 'success')
+            flash("Your password has been changed.", "success")
         else:
             time.sleep(1)
-            flash('Invalid password.', 'danger')
-            return redirect(url_for('account.change_password'))
-        return redirect(url_for('account.account'))
+            flash("Invalid password.", "danger")
+            return redirect(url_for("account.change_password"))
+        return redirect(url_for("account.account"))
 
-    return render_template('account/change_password_form.html', title='Change password', form=form)
+    return render_template("account/change_password_form.html", title="Change password", form=form)
 
 
-@bp.route('/update', methods=['GET', 'POST'])
+@bp.route("/update", methods=["GET", "POST"])
 @fresh_login_required
-@limiter.limit("1 per second", key_func=lambda: current_user.id)
+@limiter.limit("2 per second", key_func=lambda: current_user.id)
 def update():
     form = AccountForm()
     if form.validate_on_submit():
@@ -53,33 +52,43 @@ def update():
             current_user.email_address = form.email_address.data
             current_user.activated_at = None
             send_confirmation_email(current_user)
-            flash("We've sent an email to {0} with instructions to confirm your email address."
-                  .format(current_user.email_address), 'success')
+            flash(
+                "We've sent an email to {0} with instructions to confirm your email address.".format(
+                    current_user.email_address
+                ),
+                "success",
+            )
         current_user.timezone = form.timezone.data
         current_user.updated_at = pytz.utc.localize(datetime.utcnow())
         db.session.add(current_user)
         db.session.commit()
-        flash('Your account has been updated.', 'success')
-        return redirect(url_for('account.account'))
-    elif request.method == 'GET':
+        flash("Account changes have been saved.", "success")
+        return redirect(url_for("account.account"))
+    elif request.method == "GET":
         form.email_address.data = current_user.email_address
         form.timezone.data = current_user.timezone
-    return render_template('account/account_form.html', title='Update account', form=form)
+    return render_template("account/account_form.html", title="Edit account", form=form)
 
 
-@bp.route('/delete')
+@bp.route("/delete", methods=["GET", "POST"])
 @fresh_login_required
-@limiter.limit("1 per second", key_func=lambda: current_user.id)
+@limiter.limit("2 per second", key_func=lambda: current_user.id)
 def delete():
-    db.session.delete(current_user)
-    db.session.commit()
-    flash('Your account and all personal information has been permanently deleted.', 'success')
-    return redirect(url_for('main.index'))
+    if request.method == "GET":
+        return render_template("account/delete_account.html", title="Delete account")
+    elif request.method == "POST":
+        db.session.delete(current_user)
+        db.session.commit()
+        flash(
+            "Your account and all personal information has been permanently deleted.",
+            "success",
+        )
+        return redirect(url_for("main.index"))
 
 
-@bp.route('/schedule', methods=['GET', 'POST'])
+@bp.route("/schedule", methods=["GET", "POST"])
 @fresh_login_required
-@limiter.limit("1 per second", key_func=lambda: current_user.id)
+@limiter.limit("2 per second", key_func=lambda: current_user.id)
 def schedule():
     form = ScheduleForm()
     if form.validate_on_submit():
@@ -93,9 +102,9 @@ def schedule():
         current_user.updated_at = pytz.utc.localize(datetime.utcnow())
         db.session.add(current_user)
         db.session.commit()
-        flash('Your schedule has been updated', 'success')
-        return redirect(url_for('account.account'))
-    elif request.method == 'GET':
+        flash("Schedule changes have been saved", "success")
+        return redirect(url_for("account.account"))
+    elif request.method == "GET":
         mon_hours, mon_remainder = divmod(current_user.monday, 3600)
         mon_minutes, mon_seconds = divmod(mon_remainder, 60)
         form.monday.data = time(hour=mon_hours, minute=mon_minutes, second=mon_seconds)
@@ -123,4 +132,4 @@ def schedule():
         sun_hours, sun_remainder = divmod(current_user.sunday, 3600)
         sun_minutes, sun_seconds = divmod(sun_remainder, 60)
         form.sunday.data = time(hour=sun_hours, minute=sun_minutes, second=sun_seconds)
-    return render_template('account/schedule_form.html', title='Update schedule', form=form)
+    return render_template("account/schedule_form.html", title="Edit schedule", form=form)
